almost everything working

+++ Angel Martin/myNotes/src/myNotes/DatabaseManager.java
@@ -14,11 +14,11 @@ public class DatabaseManager {
 	
 	
 	public boolean addNote(Note note) {
-		return connector.addNote(note);
+		return displaySaveResult(connector.addNote(note));
 	}
 
 	public List<Note> getNotes(String toSeek) {
-		return connector.getNotes(toSeek);
+		return displayGetResult(connector.getNotes(toSeek));
 	}
 	
 	public List<Note> getNotes() {
@@ -26,7 +26,12 @@ public class DatabaseManager {
 	}
 
 	public void deleteNotes(String toSeek, boolean deleteAll) {
-		connector.deleteNotes(toSeek, deleteAll);
+		if ("".equals(toSeek) && !deleteAll) {
+			System.out.println(">>>: No text to look for...");
+			System.out.println(">>>: Do you want to delete all notes? if so, use: mynotes delete -a");
+		} else {
+			connector.deleteNotes(toSeek, deleteAll);
+		}
 	}
 
 
@@ -42,11 +47,30 @@ public class DatabaseManager {
 				System.out.println(">>>: Waiting for authentication, please wait...");
 				break;
 			case "mongorest":
-				this.connector = new MongoConnector();
+				this.connector = new MongoRESTConnector();
 				System.out.println(">>>: Using MongoDB RESP API connection...");
 				break;
 		}
 	}
 	
+	private boolean displaySaveResult(boolean saved) {
+		if (saved)
+			System.out.println(">>>: ...saved...");
+		else{
+			System.out.println(">>>: there was some problem saving your note, please check your network connectivity.     |:|  =D~~~~~");
+		}
+		
+		return saved;
+	}
+	
+	private List<Note> displayGetResult(List<Note> notes) {
+		if (notes.isEmpty()) {
+			Note auxNote = new Note("Sorry, the are currently no notes at the Data Base... :(");
+			auxNote.setId("-1");
+			notes.add(auxNote);
+		}
+		
+		return notes;
+	}
 	
 }
